+++ src/hooks/focus/useFocusTrap.ts
@@ -146,19 +146,12 @@ export type UseFocusTrapProps = {
   container: RefObject<HTMLElement>,
   active: boolean,
   initialFocus?: RefObject<HTMLElement>,
-  /**
-   * Whether to focus the first element when the initialFocus isn't provided
-   *
-   * Focuses the container instead
-   */
-  focusFirst?: boolean,
 }
 
 export const useFocusTrap = ({
   container,
   active,
   initialFocus,
-  focusFirst = true,
 }: UseFocusTrapProps) => {
   const lastFocusRef = useRef<HTMLElement>(null)
   const [paused, setPaused] = useState(false)
@@ -168,23 +161,19 @@ export const useFocusTrap = ({
     const containerElement = container.current
     // Try in the following order
     // 1. Focus the initial element
-    // 2. Focus the first focusable element in the container, focusFirst is true
-    // 3. Focus the container
+    // 2. Focus the first focusable element in the container
     if (initialFocus?.current) {
       initialFocus.current.focus()
-    } else if (!focusFirst) {
-      containerElement.focus()
     } else {
       const elements = getContainedFocusableElements(containerElement)
       if (elements && elements.length > 0) {
         const first = elements.item(0) as HTMLElement
         first.focus()
       } else {
-        console.warn('No focusable elements found in the focus trap, focusing the container instead. Make sure this is the correct behavior. Affected element: ', containerElement)
-        containerElement.focus()
+        console.warn('No focusable elements found in the focus trap. Affected element: ', containerElement)
       }
     }
-  }, [container, focusFirst, initialFocus])
+  }, [container, initialFocus])
 
   useEffect(() => {
     if (active) {